+++ villagesql/sdk/include/villagesql/abi/types.h
@@ -21,7 +21,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "storage.h"
+#include "preview/storage.h"
 
 // Protocol Versioning
 //
